+++ src/pages/Dashboard/functions/calculatePoints.js
